requirements.txt
Django==2.2.13
djangorestframework==3.11.0
django_environ==0.4.5
psycopg2==2.8.5
drf-yasg==1.17.1
django-redis==4.12.1
